Improved error handling

+++ scripts/explorer.sh
@@ -7,7 +7,7 @@ start_explorer() {
     (stop_explorer) &
     __spinner
 
-    if [[ ! $(docker ps | grep fabric) ]]; then
+    if ! docker ps | grep -q "fabric"; then
         logerr "No Fabric networks running. First launch fabkit start"
         exit 1
     fi
@@ -21,15 +21,23 @@ start_explorer() {
     config="${FABKIT_EXPLORER_PATH}/connection-profile/first-network"
     admin_key_path="peerOrganizations/org1.example.com/users/Admin@org1.example.com/msp/keystore"
     private_key="/tmp/crypto/${admin_key_path}/$(ls ${FABKIT_CRYPTOS_PATH}/${admin_key_path})"
-    __jq <"${config}.base.json" -r --arg private_key "$private_key" '.organizations.Org1MSP.adminPrivateKey.path = $private_key' |
-        __jq -r --argjson FABKIT_TLS_ENABLED "$FABKIT_TLS_ENABLED" '.client.tlsEnable = $FABKIT_TLS_ENABLED' >${config}.json || exit 1
+    if (__jq <"${config}.base.json" -r --arg private_key "$private_key" '.organizations.Org1MSP.adminPrivateKey.path = $private_key' |
+        __jq -r --argjson FABKIT_TLS_ENABLED "$FABKIT_TLS_ENABLED" '.client.tlsEnable = $FABKIT_TLS_ENABLED' >"${config}.json") 2>&1 >/dev/null | grep -iE "erro|pani|fail|fatal" > >(__throw >&2); then
+        logerr "Error in replacing private key in explorer configuration"
+        exit 1
+    fi
 
     # considering tls enabled as default in base
     if [ "${FABKIT_TLS_ENABLED:-}" = "false" ]; then
-        sed -i'.bak' -e 's/grpcs/grpc/g' -e 's/https/http/g' ${config}.json && rm ${config}.json.bak 1>/dev/null 2> >(__throw >&2)
+        sed -i'.bak' -e 's/grpcs/grpc/g' -e 's/https/http/g' "${config}.json" && rm "${config}.json.bak" 1>/dev/null 2> >(__throw >&2)
     fi
 
-    (loginfo "Launching containers" && docker-compose --env-file "${FABKIT_ROOT}/.env" -f "${FABKIT_EXPLORER_PATH}/docker-compose.yaml" up --force-recreate -d &>/dev/null) &
+    (
+        if loginfo "Launching containers" && docker-compose --env-file "${FABKIT_ROOT}/.env" -f "${FABKIT_EXPLORER_PATH}/docker-compose.yaml" up --force-recreate -d 2>&1 >/dev/null | grep -iE "erro|pani|fail|fatal" > >(__throw >&2); then
+            logerr "Failed to launch containers"
+            exit 1
+        fi
+    ) &
     __spinner
 
     echo "Blockchain Explorer default user is exploreradmin/exploreradminpw - $(logsucc http://localhost:8090)"
@@ -39,6 +47,9 @@ start_explorer() {
 stop_explorer() {
     loginfo "Stopping explorer"
 
-    docker-compose --env-file "${FABKIT_ROOT}/.env" -f "${FABKIT_EXPLORER_PATH}/docker-compose.yaml" down &>/dev/null || exit 1
+    if docker-compose --env-file "${FABKIT_ROOT}/.env" -f "${FABKIT_EXPLORER_PATH}/docker-compose.yaml" down 2>&1 >/dev/null | grep -iE "erro|pani|fail|fatal" > >(__throw >&2); then
+        logerr "Failed to stop containers"
+        exit 1
+    fi
     docker volume prune -f $(docker volume ls | awk '($2 ~ /explorer/) {print $2}') &>/dev/null || true
 }
